fix(gold): partner-dashboard 500(ROUND 캐스팅) + 골드문의 지사장 범위(본인→지사 전체)
①partner-dashboard/gold 500: $queryRaw가 ROUND(AVG(float)*100, 2) 호출 → Postgres엔 round(double precision,int) 없음(42883) → ROUND((...)::numeric, 2) 캐스팅. /partner-dashboard·/marketing/sales 골드회원탭 같은 엔드포인트라 둘 다 해결.

②골드문의 가시성: OWNER가 '본인이 manager/agent인 문의만'으로 좁던 것 → 사장님 모델대로 '지사 전체(organizationId)=본인+산하 대리점장·마케터'로 확대(다른 지사 불가). OWNER는 organizationId 기준이라 mallUser 불필요(403가드는 AGENT만). 위험한 '!mallUser→empty(전체노출)' 폴백 제거(미충족시 1=0 차단). AGENT=본인 agentId+조직격리 유지, GLOBAL_ADMIN=전체.

tsc 0. Co-Authored-By: Claude Opus 4.8 (1M context) <noreply@anthropic.com>

+++ src/app/api/partner/dashboard/gold/route.ts
@@ -94,7 +94,7 @@ export async function GET(req: Request) {
     if (isAdmin) {
       const result = await prisma.$queryRaw<Array<{ rate: number | null }>>`
         SELECT CASE WHEN COUNT(*) = 0 THEN 0
-          ELSE ROUND(AVG("paidCount"::float / NULLIF("totalPayments", 0)) * 100, 2)
+          ELSE ROUND((AVG("paidCount"::float / NULLIF("totalPayments", 0)) * 100)::numeric, 2)
         END AS rate
         FROM "GoldMember"
         WHERE "status" = 'ACTIVE' AND "totalPayments" > 0
@@ -106,7 +106,7 @@ export async function GET(req: Request) {
     } else if (orgId) {
       const result = await prisma.$queryRaw<Array<{ rate: number | null }>>`
         SELECT CASE WHEN COUNT(*) = 0 THEN 0
-          ELSE ROUND(AVG("paidCount"::float / NULLIF("totalPayments", 0)) * 100, 2)
+          ELSE ROUND((AVG("paidCount"::float / NULLIF("totalPayments", 0)) * 100)::numeric, 2)
         END AS rate
         FROM "GoldMember"
         WHERE "status" = 'ACTIVE' AND "totalPayments" > 0
